model-vs-human-training/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
idna==3.3
joblib==1.1.0
numpy==1.21.4
pandas==1.3.4
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.0.0
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==4.0.0
urllib3==1.26.7
